requirements.txt
pytest
pylint
numpy
plotly
kaleido
jupyter
pandas
black[jupyter]
Pillow
